backend/requirements.txt
# Core Django & ASGI
Django>=4.2.0
daphne>=4.0.0
channels>=4.0.0
channels-redis>=4.1.0

# Database & ORM
psycopg2-binary>=2.9.0
django-extensions>=3.2.0

# API & Serialization
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
graphene-django>=3.0.0

# Background Tasks
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.4.0

# HTTP Client & API
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.0

# AI/ML & Embeddings
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# Ollama Integration
ollama>=0.1.7

# OpenRouter/OpenAI Integration
openai>=1.0.0

# Data Processing
pandas>=2.0.0
python-dateutil>=2.8.0

# Time-series Database
influxdb-client>=1.40.0

# Caching & Performance
django-cache-machine>=1.2.0
django-redis>=5.2.0

# Monitoring & Logging
django-health-check>=3.17.0
sentry-sdk>=1.25.0

# Environment & Config
python-decouple>=3.8
django-environ>=0.10.0

# Development
pytest>=7.3.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0